chore(review): fixed issues by Alexey
