+++ .eslintrc.js
@@ -27,6 +27,7 @@ module.exports = {
     '@typescript-eslint/no-var-requires': 'off',
     '@typescript-eslint/deprecation': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
+    "jsx-quotes": ["error", "prefer-double"],
 
     'comma-dangle': 'off',
     'space-before-function-paren': 'off',
