+++ package.json
@@ -21,7 +21,6 @@
     "@tanstack/react-query": "^5.8.1",
     "axios": "^1.6.0",
     "effector": "^22.8.7",
-    "effector-localstorage": "^1.0.0",
     "effector-react": "^22.5.4",
     "effector-storage": "^6.1.1",
     "react": "18.2.0",
