+++ src/pages/auth/otp-code/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useOtpCode } from './use-otp-code'
