+++ src/pages/auth/password-writing/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useKeyboardAnimation } from './use-keyboard-animation'
