+++ src/shared/ui/templates/keyboard-template/index.ts
@@ -1,0 +1,1 @@
+export { KeyboardTemplate } from './keyboard-template'
