+++ src/shared/ui/templates/keyboard-template/types.ts
@@ -6,8 +6,3 @@ export type TKeyboardButton = {
     type: TKeyboardButtonType
     value?: number
 }
-
-export type TimerInfo = {
-    minutes: number
-    seconds: number
-}
